+++ Projects/Claims/Sources/Views/SubmitClaim/SubViews/ClaimContactCard.swift
@@ -1,13 +1,11 @@
 import Combine
 import Foundation
 import Kingfisher
-import PresentableStore
 import SwiftUI
 import hCore
 import hCoreUI
 
 struct ClaimContactCard: View {
-    @PresentableStore var store: SubmitClaimStore
     var model: FlowClaimDeflectStepModel
     var title: String?
     var imageUrl: String?
@@ -102,7 +100,6 @@ struct ClaimContactCard: View {
 }
 
 struct ClaimEmergencyContactCard: View {
-    @PresentableStore var store: SubmitClaimStore
     let cardTitle: String?
     let footnote: String?
     let imageUrl: String?
